Refactor destroy commands
