+++ decidim-participatory_processes/app/commands/decidim/participatory_processes/admin/destroy_participatory_process_step.rb
@@ -4,47 +4,16 @@ module Decidim
   module ParticipatoryProcesses
     module Admin
       # A command that sets all steps in a participatory process as inactive
-      class DestroyParticipatoryProcessStep < Decidim::Command
-        # Public: Initializes the command.
-        #
-        # step - A ParticipatoryProcessStep that will be deactivated
-        # current_user - the user performing the action
-        def initialize(step, current_user)
-          @step = step
-          @participatory_process = step.participatory_process
-          @current_user = current_user
-        end
-
-        # Executes the command. Broadcasts these events:
-        #
-        # - :ok when everything is valid.
-        # - :invalid if the data was not valid and we could not proceed.
-        #
-        # Returns nothing.
-        def call
-          return broadcast(:invalid, :active_step) if active_step?
-
-          transaction do
-            Decidim.traceability.perform_action!(:delete, step, current_user) do
-              step.destroy!
-              step
-            end
-
-            reorder_steps
-          end
-
-          broadcast(:ok)
-        end
-
+      class DestroyParticipatoryProcessStep < Decidim::Commands::DestroyResource
         private
 
-        attr_reader :step, :participatory_process, :current_user
+        def participatory_process = resource.participatory_space
 
-        def active_step?
-          participatory_process.steps.count > 1 && step.active?
+        def invalid?
+          participatory_process.steps.count > 1 && resource.active?
         end
 
-        def reorder_steps
+        def run_after_hooks
           steps = participatory_process.steps.reload
 
           ReorderParticipatoryProcessSteps
